Enable PHPStan in tests

+++ lib/Doctrine/Migrations/Tools/BytesFormatter.php
@@ -22,6 +22,6 @@ final class BytesFormatter
         $base     = log($size, 1024);
         $suffixes = ['', 'K', 'M', 'G', 'T'];
 
-        return round(pow(1024, $base - floor($base)), $precision) . $suffixes[floor($base)];
+        return round(pow(1024, $base - floor($base)), $precision) . $suffixes[(int) floor($base)];
     }
 }
